feat: AbtractDecoder and LavalinkDecoder class
This commit will add 2 new class is AbstractDecoder and LavalinkDecoder class. AdstractDecoder is for making custom decoder. LavalinkDecoder is an example of AbstractDecoder
