+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rainlink",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "Another lavalink wrapper but focus on stability and rich features",
   "repository": {
     "type": "git",
